backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18
Pillow==11.1.0
opencv-python==4.10.0.84
pytesseract==0.3.13
psycopg2-binary==2.9.10
sqlalchemy==2.0.36
python-dotenv==1.0.1
httpx==0.28.1
python-dateutil==2.9.0.post0
google-generativeai==0.8.6
